Start build.cjs
