+++ gulpfile.cjs
@@ -14,12 +14,10 @@ const globby = require("globby");
 const gulpTap = require("gulp-tap");
 const open = require("open");
 const rimraf = require("rimraf");
-const glslStripComments = require("glsl-strip-comments");
 const mkdirp = require("mkdirp");
 const mergeStream = require("merge-stream");
 const streamToPromise = require("stream-to-promise");
 const gulp = require("gulp");
-const gulpInsert = require("gulp-insert");
 const gulpZip = require("gulp-zip");
 const gulpRename = require("gulp-rename");
 const gulpReplace = require("gulp-replace");
@@ -28,16 +26,22 @@ const Karma = require("karma");
 const yargs = require("yargs");
 const AWS = require("aws-sdk");
 const mime = require("mime");
-const cleanCSS = require("gulp-clean-css");
 const typescript = require("typescript");
 const esbuild = require("esbuild");
 const istanbul = require("istanbul-lib-instrument");
-const rollup = require("rollup");
-const rollupPluginStripPragma = require("rollup-plugin-strip-pragma");
-const rollupPluginTerser = require("rollup-plugin-terser");
-const rollupCommonjs = require("@rollup/plugin-commonjs");
-const rollupResolve = require("@rollup/plugin-node-resolve").default;
 
+const {
+  createCesiumJs,
+  buildCesiumJs,
+  buildWorkers,
+  glslToJavaScript,
+  createSpecList,
+  buildSpecs,
+  createGalleryList,
+  createJsHintOptions,
+  esbuildBaseConfig,
+  stripPragmaPlugin,
+} = require("./build.cjs");
 const packageJson = require("./package.json");
 let version = packageJson.version;
 if (/\.0$/.test(version)) {
@@ -93,64 +97,10 @@ const watchedSpecFiles = [
   "Specs/*.js",
   "Specs/TestWorkers/*.js",
 ];
-const cssFiles = "Source/**/*.css";
 const shaderFiles = [
   "Source/Shaders/**/*.glsl",
   "Source/ThirdParty/Shaders/*.glsl",
 ];
-
-let copyrightHeader = fs.readFileSync(
-  path.join("Source", "copyrightHeader.js"),
-  "utf8"
-);
-copyrightHeader = copyrightHeader.replace("${version}", version);
-
-function escapeCharacters(token) {
-  return token.replace(/[\-\[\]\/\{\}\(\)\*\+\?\.\\\^\$\|]/g, "\\$&");
-}
-
-function constructRegex(pragma, exclusive) {
-  const prefix = exclusive ? "exclude" : "include";
-  pragma = escapeCharacters(pragma);
-
-  const s =
-    `[\\t ]*\\/\\/>>\\s?${prefix}Start\\s?\\(\\s?(["'])${pragma}\\1\\s?,\\s?pragmas\\.${pragma}\\s?\\)\\s?;?` +
-    // multiline code block
-    `[\\s\\S]*?` +
-    // end comment
-    `[\\t ]*\\/\\/>>\\s?${prefix}End\\s?\\(\\s?(["'])${pragma}\\2\\s?\\)\\s?;?\\s?[\\t ]*\\n?`;
-
-  return new RegExp(s, "gm");
-}
-
-const pragmas = {
-  debug: false,
-};
-const stripPragmaPlugin = {
-  name: "strip-pragmas",
-  setup: (build) => {
-    const readFile = Promise.promisify(fs.readFile);
-    build.onLoad({ filter: /\.js$/ }, async (args) => {
-      let source = await readFile(args.path, "utf8");
-
-      try {
-        for (const key in pragmas) {
-          if (pragmas.hasOwnProperty(key)) {
-            source = source.replace(constructRegex(key, pragmas[key]), "");
-          }
-        }
-
-        return { contents: source };
-      } catch (e) {
-        return {
-          errors: {
-            text: e.message,
-          },
-        };
-      }
-    });
-  },
-};
 
 // Print an esbuild warning
 function printBuildWarning({ location, text }) {
@@ -178,200 +128,6 @@ function handleBuildWarnings(result) {
       printBuildWarning(warning);
     }
   }
-}
-
-const esbuildBaseConfig = {
-  target: "es2020",
-  legalComments: "inline",
-  banner: {
-    js: copyrightHeader,
-  },
-};
-
-async function buildCesiumJs(options) {
-  const css = globby.sync(cssFiles);
-
-  const buildConfig = {
-    ...esbuildBaseConfig,
-    entryPoints: ["Source/Cesium.js"],
-    bundle: true,
-    minify: options.minify,
-    sourcemap: options.sourcemap,
-    external: ["https", "http", "url", "zlib"],
-    plugins: options.removePragmas ? [stripPragmaPlugin] : undefined,
-    incremental: options.incremental,
-    logLevel: "error", // print errors immediately, and collect warnings so we can filter out known ones
-  };
-
-  // Build ESM
-  const result = await esbuild.build({
-    ...buildConfig,
-    format: "esm",
-    outfile: path.join(options.path, "index.js"),
-  });
-
-  handleBuildWarnings(result);
-
-  const results = [result];
-
-  // Copy and minify CSS and third party
-  await esbuild.build({
-    ...esbuildBaseConfig,
-    entryPoints: [
-      "Source/ThirdParty/google-earth-dbroot-parser.js",
-      ...css, // Load and optionally minify css
-    ],
-    loader: {
-      ".gif": "text",
-      ".png": "text",
-    },
-    minify: options.minify,
-    sourcemap: options.sourcemap,
-    outdir: options.path,
-  });
-
-  // Build IIFE
-  if (options.iife) {
-    const result = await esbuild.build({
-      ...buildConfig,
-      format: "iife",
-      globalName: "Cesium",
-      outfile: path.join(options.path, "Cesium.js"),
-    });
-
-    handleBuildWarnings(result);
-
-    results.push(result);
-  }
-
-  if (options.node) {
-    const result = await esbuild.build({
-      ...buildConfig,
-      format: "cjs",
-      platform: "node",
-      sourcemap: false,
-      outfile: path.join(options.path, "index.cjs"),
-    });
-
-    handleBuildWarnings(result);
-    results.push(result);
-  }
-
-  return results;
-}
-
-function rollupWarning(message) {
-  // Ignore eval warnings in third-party code we don't have control over
-  if (message.code === "EVAL" && /protobufjs/.test(message.loc.file)) {
-    return;
-  }
-
-  console.log(message);
-}
-
-/**
- * Bundles the workers and outputs the result to the specified directory
- * @param {Object} options
- * @param {boolean} [options.minify=false] true if the worker output should be minified
- * @param {boolean} [options.removePragmas=false] true if debug pragma should be removed
- * @param {boolean} [options.sourcemap=false] true if an external sourcemap should be generated
- * @param {String} options.path output directory
- */
-async function buildWorkers(options) {
-  // Copy existing workers
-  const workers = globby.sync([
-    "Source/Workers/**",
-    "Source/ThirdParty/Workers/**",
-  ]);
-
-  await esbuild.build({
-    ...esbuildBaseConfig,
-    entryPoints: workers,
-    outdir: options.path,
-    outbase: "Source", // Maintain existing file paths
-    minify: options.minify,
-  });
-
-  // Use rollup to build the workers:
-  // 1) They can be built as AMD style modules
-  // 2) They can be built using code-splitting, resulting in smaller modules
-  return globby(["Source/WorkersES6/*.js"]).then(function (files) {
-    const plugins = [rollupResolve(), rollupCommonjs()];
-
-    if (options.removePragmas) {
-      plugins.push(
-        rollupPluginStripPragma({
-          pragmas: ["debug"],
-        })
-      );
-    }
-
-    if (options.minify) {
-      plugins.push(rollupPluginTerser.terser());
-    }
-
-    return rollup
-      .rollup({
-        input: files,
-        plugins: plugins,
-        onwarn: rollupWarning,
-      })
-      .then(function (bundle) {
-        return bundle.write({
-          dir: path.join(options.path, "Workers"),
-          format: "amd",
-          // Rollup cannot generate a sourcemap
-          sourcemap: options.sourcemap && !options.removePragmas,
-          banner: copyrightHeader,
-        });
-      });
-  });
-}
-
-const externalResolvePlugin = {
-  name: "external-cesium",
-  setup: (build) => {
-    build.onResolve({ filter: new RegExp(`Cesium\.js$`) }, () => {
-      return {
-        path: "Cesium",
-        namespace: "external-cesium",
-      };
-    });
-
-    build.onLoad(
-      {
-        filter: new RegExp(`^Cesium$`),
-        namespace: "external-cesium",
-      },
-      () => {
-        const contents = `module.exports = Cesium`;
-        return {
-          contents,
-        };
-      }
-    );
-  },
-};
-
-async function buildSpecs(options) {
-  options = options || {};
-
-  const results = await esbuild.build({
-    entryPoints: [
-      "Specs/spec-main.js",
-      "Specs/SpecList.js",
-      "Specs/karma-main.js",
-    ],
-    bundle: true,
-    format: "esm",
-    sourcemap: true,
-    target: "es2020",
-    outdir: path.join("Build", "Specs"),
-    plugins: [externalResolvePlugin],
-    incremental: options.incremental,
-  });
-
-  return results;
 }
 
 async function build(options) {
@@ -411,7 +167,7 @@ async function build(options) {
       path: outputDirectory,
       removePragmas: options.removePragmas,
     }),
-    createGalleryList(),
+    createGalleryList(noDevelopmentGallery),
     buildSpecs()
   ).then(() => {
     return copyAssets({
@@ -1426,145 +1182,6 @@ function copyAssets(options) {
   return streamToPromise(stream);
 }
 
-function glslToJavaScript(minify, minifyStateFilePath) {
-  fs.writeFileSync(minifyStateFilePath, minify.toString());
-  const minifyStateFileLastModified = fs.existsSync(minifyStateFilePath)
-    ? fs.statSync(minifyStateFilePath).mtime.getTime()
-    : 0;
-
-  // collect all currently existing JS files into a set, later we will remove the ones
-  // we still are using from the set, then delete any files remaining in the set.
-  const leftOverJsFiles = {};
-
-  globby
-    .sync(["Source/Shaders/**/*.js", "Source/ThirdParty/Shaders/*.js"])
-    .forEach(function (file) {
-      leftOverJsFiles[path.normalize(file)] = true;
-    });
-
-  const builtinFunctions = [];
-  const builtinConstants = [];
-  const builtinStructs = [];
-
-  const glslFiles = globby.sync(shaderFiles);
-  glslFiles.forEach(function (glslFile) {
-    glslFile = path.normalize(glslFile);
-    const baseName = path.basename(glslFile, ".glsl");
-    const jsFile = `${path.join(path.dirname(glslFile), baseName)}.js`;
-
-    // identify built in functions, structs, and constants
-    const baseDir = path.join("Source", "Shaders", "Builtin");
-    if (
-      glslFile.indexOf(path.normalize(path.join(baseDir, "Functions"))) === 0
-    ) {
-      builtinFunctions.push(baseName);
-    } else if (
-      glslFile.indexOf(path.normalize(path.join(baseDir, "Constants"))) === 0
-    ) {
-      builtinConstants.push(baseName);
-    } else if (
-      glslFile.indexOf(path.normalize(path.join(baseDir, "Structs"))) === 0
-    ) {
-      builtinStructs.push(baseName);
-    }
-
-    delete leftOverJsFiles[jsFile];
-
-    const jsFileExists = fs.existsSync(jsFile);
-    const jsFileModified = jsFileExists
-      ? fs.statSync(jsFile).mtime.getTime()
-      : 0;
-    const glslFileModified = fs.statSync(glslFile).mtime.getTime();
-
-    if (
-      jsFileExists &&
-      jsFileModified > glslFileModified &&
-      jsFileModified > minifyStateFileLastModified
-    ) {
-      return;
-    }
-
-    let contents = fs.readFileSync(glslFile, "utf8");
-    contents = contents.replace(/\r\n/gm, "\n");
-
-    let copyrightComments = "";
-    const extractedCopyrightComments = contents.match(
-      /\/\*\*(?:[^*\/]|\*(?!\/)|\n)*?@license(?:.|\n)*?\*\//gm
-    );
-    if (extractedCopyrightComments) {
-      copyrightComments = `${extractedCopyrightComments.join("\n")}\n`;
-    }
-
-    if (minify) {
-      contents = glslStripComments(contents);
-      contents = contents
-        .replace(/\s+$/gm, "")
-        .replace(/^\s+/gm, "")
-        .replace(/\n+/gm, "\n");
-      contents += "\n";
-    }
-
-    contents = contents.split('"').join('\\"').replace(/\n/gm, "\\n\\\n");
-    contents = `${copyrightComments}\
-//This file is automatically rebuilt by the Cesium build process.\n\
-export default "${contents}";\n`;
-
-    fs.writeFileSync(jsFile, contents);
-  });
-
-  // delete any left over JS files from old shaders
-  Object.keys(leftOverJsFiles).forEach(function (filepath) {
-    rimraf.sync(filepath);
-  });
-
-  const generateBuiltinContents = function (contents, builtins, path) {
-    for (let i = 0; i < builtins.length; i++) {
-      const builtin = builtins[i];
-      contents.imports.push(
-        `import czm_${builtin} from './${path}/${builtin}.js'`
-      );
-      contents.builtinLookup.push(`czm_${builtin} : ` + `czm_${builtin}`);
-    }
-  };
-
-  //generate the JS file for Built-in GLSL Functions, Structs, and Constants
-  const contents = {
-    imports: [],
-    builtinLookup: [],
-  };
-  generateBuiltinContents(contents, builtinConstants, "Constants");
-  generateBuiltinContents(contents, builtinStructs, "Structs");
-  generateBuiltinContents(contents, builtinFunctions, "Functions");
-
-  const fileContents = `//This file is automatically rebuilt by the Cesium build process.\n${contents.imports.join(
-    "\n"
-  )}\n\nexport default {\n    ${contents.builtinLookup.join(",\n    ")}\n};\n`;
-
-  fs.writeFileSync(
-    path.join("Source", "Shaders", "Builtin", "CzmBuiltins.js"),
-    fileContents
-  );
-}
-
-function createCesiumJs() {
-  let contents = `export const VERSION = '${version}';\n`;
-  globby.sync(sourceFiles).forEach(function (file) {
-    file = path.relative("Source", file);
-
-    let moduleId = file;
-    moduleId = filePathToModuleId(moduleId);
-
-    let assignmentName = path.basename(file, path.extname(file));
-    if (moduleId.indexOf("Shaders/") === 0) {
-      assignmentName = `_shaders${assignmentName}`;
-    }
-    assignmentName = assignmentName.replace(/(\.|-)/g, "_");
-    contents += `export { default as ${assignmentName} } from './${moduleId}.js';${os.EOL}`;
-  });
-
-  fs.writeFileSync("Source/Cesium.js", contents);
-}
-
 function createTypeScriptDefinitions() {
   // Run jsdoc with tsd-jsdoc to generate an initial Cesium.d.ts file.
   child_process.execSync("npx jsdoc --configure Tools/jsdoc/ts-conf.json", {
@@ -1714,20 +1331,6 @@ ${source}
   }
 }
 
-function createSpecList() {
-  const files = globby.sync(["Specs/**/*Spec.js"]);
-
-  let contents = "";
-  files.forEach(function (file) {
-    contents += `import './${filePathToModuleId(file).replace(
-      "Specs/",
-      ""
-    )}.js';\n`;
-  });
-
-  fs.writeFileSync(path.join("Specs", "SpecList.js"), contents);
-}
-
 /**
  * Reads `ThirdParty.extra.json` file
  * @param path {string} Path to `ThirdParty.extra.json`
@@ -1860,115 +1463,6 @@ function generateThirdParty() {
         JSON.stringify(licenseJson, null, 2)
       );
     });
-}
-
-function createGalleryList() {
-  const demoObjects = [];
-  const demoJSONs = [];
-  const output = path.join("Apps", "Sandcastle", "gallery", "gallery-index.js");
-
-  const fileList = ["Apps/Sandcastle/gallery/**/*.html"];
-  if (noDevelopmentGallery) {
-    fileList.push("!Apps/Sandcastle/gallery/development/**/*.html");
-  }
-
-  // On travis, the version is set to something like '1.43.0-branch-name-travisBuildNumber'
-  // We need to extract just the Major.Minor version
-  const majorMinor = packageJson.version.match(/^(.*)\.(.*)\./);
-  const major = majorMinor[1];
-  const minor = Number(majorMinor[2]) - 1; // We want the last release, not current release
-  const tagVersion = `${major}.${minor}`;
-
-  // Get an array of demos that were added since the last release.
-  // This includes newly staged local demos as well.
-  let newDemos = [];
-  try {
-    newDemos = child_process
-      .execSync(
-        `git diff --name-only --diff-filter=A ${tagVersion} Apps/Sandcastle/gallery/*.html`,
-        { stdio: ["pipe", "pipe", "ignore"] }
-      )
-      .toString()
-      .trim()
-      .split("\n");
-  } catch (e) {
-    // On a Cesium fork, tags don't exist so we can't generate the list.
-  }
-
-  let helloWorld;
-  globby.sync(fileList).forEach(function (file) {
-    const demo = filePathToModuleId(
-      path.relative("Apps/Sandcastle/gallery", file)
-    );
-
-    const demoObject = {
-      name: demo,
-      isNew: newDemos.includes(file),
-    };
-
-    if (fs.existsSync(`${file.replace(".html", "")}.jpg`)) {
-      demoObject.img = `${demo}.jpg`;
-    }
-
-    demoObjects.push(demoObject);
-
-    if (demo === "Hello World") {
-      helloWorld = demoObject;
-    }
-  });
-
-  demoObjects.sort(function (a, b) {
-    if (a.name < b.name) {
-      return -1;
-    } else if (a.name > b.name) {
-      return 1;
-    }
-    return 0;
-  });
-
-  const helloWorldIndex = Math.max(demoObjects.indexOf(helloWorld), 0);
-
-  for (let i = 0; i < demoObjects.length; ++i) {
-    demoJSONs[i] = JSON.stringify(demoObjects[i], null, 2);
-  }
-
-  const contents = `\
-// This file is automatically rebuilt by the Cesium build process.\n\
-const hello_world_index = ${helloWorldIndex};\n\
-const VERSION = '${version}';\n\
-const gallery_demos = [${demoJSONs.join(", ")}];\n\
-const has_new_gallery_demos = ${newDemos.length > 0 ? "true;" : "false;"}\n`;
-
-  fs.writeFileSync(output, contents);
-
-  // Compile CSS for Sandcastle
-  return streamToPromise(
-    gulp
-      .src(path.join("Apps", "Sandcastle", "templates", "bucketRaw.css"))
-      .pipe(cleanCSS())
-      .pipe(gulpRename("bucket.css"))
-      .pipe(
-        gulpInsert.prepend(
-          "/* This file is automatically rebuilt by the Cesium build process. */\n"
-        )
-      )
-      .pipe(gulp.dest(path.join("Apps", "Sandcastle", "templates")))
-  );
-}
-
-function createJsHintOptions() {
-  const jshintrc = JSON.parse(
-    fs.readFileSync(path.join("Apps", "Sandcastle", ".jshintrc"), "utf8")
-  );
-
-  const contents = `\
-// This file is automatically rebuilt by the Cesium build process.\n\
-const sandcastleJsHintOptions = ${JSON.stringify(jshintrc, null, 4)};\n`;
-
-  fs.writeFileSync(
-    path.join("Apps", "Sandcastle", "jsHintOptions.js"),
-    contents
-  );
 }
 
 function buildSandcastle() {
